requirements.txt
streamlit
pandas
PyPDF2
openai
toml
mysql-connector-python
PyJWT